move `images` to demo + simplify `images` example
