+++ crates/egui_demo_app/Cargo.toml
@@ -16,9 +16,10 @@ crate-type = ["cdylib", "rlib"]
 
 
 [features]
-default = ["glow", "persistence"]
+default = ["glow", "persistence", "image_viewer"]
 
 http = ["ehttp", "image", "poll-promise", "egui_extras/image"]
+image_viewer = ["image", "egui_extras/all-loaders", "rfd"]
 persistence = ["eframe/persistence", "egui/persistence", "serde"]
 web_screen_reader = ["eframe/web_screen_reader"]                  # experimental
 serde = ["dep:serde", "egui_demo_lib/serde", "egui/serde"]
@@ -26,7 +27,6 @@ syntax_highlighting = ["egui_demo_lib/syntax_highlighting"]
 
 glow = ["eframe/glow"]
 wgpu = ["eframe/wgpu", "bytemuck"]
-
 
 [dependencies]
 chrono = { version = "0.4", default-features = false, features = [
@@ -48,6 +48,7 @@ bytemuck = { version = "1.7.1", optional = true }
 egui_extras = { version = "0.22.0", optional = true, path = "../egui_extras", features = [
   "log",
 ] }
+rfd = { version = "0.11", optional = true }
 
 # feature "http":
 ehttp = { version = "0.3.0", optional = true }
